+++ blueprints/flexberry-group/index.js
@@ -92,59 +92,17 @@ var GroupBlueprint = (function () {
         if (!fs.existsSync(metadataSubDir))
             return;
         var list = fs.readdirSync(metadataSubDir);
-        var _loop_1 = function(file) {
-            var pp = path.parse(file);
-            if (pp.ext != ".json")
-                return "continue";
-            var entityName = pp.name;
-            if (notOverwrite && fs.existsSync(folderJsFiles + "/" + entityName + ".js"))
-                return "continue";
-            var groupOptions = lodash.merge({}, this_1.options, { entity: { name: entityName } });
-            GroupBlueprint.groupOptions.push(groupOptions);
-            this_1.promise = this_1.promise.then(GroupBlueprint.funCallback).then(function () {
-                if (!(this.options.project.pkg.keywords && this.options.project.pkg.keywords["0"] === "ember-addon")) {
-                    return;
-                }
-                var middlePaths;
-                switch (this.blueprintName) {
-                    case 'flexberry-object':
-                        middlePaths = ["object", "transform"];
-                        break;
-                    case 'flexberry-enum':
-                        middlePaths = ["enum", "transform"];
-                        break;
-                    case 'flexberry-list-form':
-                        middlePaths = ["controller", "route"];
-                        break;
-                    case 'flexberry-edit-form':
-                        middlePaths = ["controller", "route"];
-                        break;
-                    case 'flexberry-model':
-                        middlePaths = ["model", "serializer"];
-                        break;
-                    default:
-                        return;
-                }
-                var promises = [];
-                for (var _i = 0, middlePaths_1 = middlePaths; _i < middlePaths_1.length; _i++) {
-                    var middlePath = middlePaths_1[_i];
-                    var flexberryAddon = Blueprint.lookup("flexberry-addon", {
-                        ui: undefined,
-                        analytics: undefined,
-                        project: undefined,
-                        paths: ["node_modules/ember-flexberry/blueprints"]
-                    });
-                    var addonBlueprintOptions = lodash.merge({}, groupOptions, { installingAddon: true, middlePath: middlePath, originBlueprintName: middlePath });
-                    promises.push(flexberryAddon["install"](addonBlueprintOptions));
-                }
-                return Promise.all(promises);
-            }.bind(this_1));
-        };
-        var this_1 = this;
         for (var _i = 0, list_1 = list; _i < list_1.length; _i++) {
             var file = list_1[_i];
-            var state_1 = _loop_1(file);
-            if (state_1 === "continue") continue;
+            var pp = path.parse(file);
+            if (pp.ext != ".json")
+                continue;
+            var entityName = pp.name;
+            if (notOverwrite && fs.existsSync(folderJsFiles + "/" + entityName + ".js"))
+                continue;
+            var groupOptions = lodash.merge({}, this.options, { entity: { name: entityName } });
+            GroupBlueprint.groupOptions.push(groupOptions);
+            this.promise = this.promise.then(GroupBlueprint.funCallback);
         }
     };
     GroupBlueprint.groupOptions = [];