+++ blueprints/flexberry-model/index.js
@@ -4,6 +4,8 @@
 "use strict";
 var ModelBlueprint_1 = require('./ModelBlueprint');
 var lodash = require('lodash');
+var Blueprint = require('ember-cli/lib/models/blueprint');
+var AddonBlueprint_1 = require('../flexberry-addon/AddonBlueprint');
 module.exports = {
     description: 'Generates an ember-data model for flexberry.',
     availableOptions: [
@@ -12,6 +14,11 @@ module.exports = {
     ],
     supportsAddon: function () {
         return false;
+    },
+    afterInstall: function (options) {
+        if (this.project.isEmberCLIAddon()) {
+            AddonBlueprint_1.default.install(options, ["model", "serializer"]);
+        }
     },
     /**
      * Blueprint Hook locals.